final enhancements
