+++ api/db/db.js
@@ -1,9 +1,3 @@
-const todos =  [
-    {
-        id: 1,
-        title: "lunch112",
-        description: "Go for lunc by 2pm"
-    }
-];
 
-export default todos;
+
+export default db;
