--- api/db/setup/package.json
@@ -1,13 +1,0 @@
-{
-  "name": "setup",
-  "version": "1.0.0",
-  "main": "index.js",
-  "license": "MIT",
-  "dependencies": {
-    "aws-sdk": "^2.643.0",
-    "dynamodb-geo": "^0.4.0"
-  },
-  "scripts": {
-    "start": "node index.js"
-  }
-}
