+++ api/db/tables/load_data_places.js
